+++ src/main/java/pl/wroc/pwr/car/CarRankComparator.java
@@ -8,7 +8,7 @@ import java.util.Comparator;
 public class CarRankComparator implements Comparator<CarRank> {
     @Override
     public int compare(CarRank o1, CarRank o2) {
-        Double rank = o1.getRank() - o2.getRank();
+        Integer rank = o1.getRank() - o2.getRank();
         if (rank < 0)
             return 1;
         else if (rank > 0) return -1;
